Improve herb patch info output and refactor some code
