+++ src/utils/mod.rs
@@ -1,4 +1,6 @@
 pub mod context;
+pub mod diary;
+pub mod farm;
 pub mod fmt;
 pub mod hiscore;
 pub mod math;
